Update communication unit and capabilities for HTTP, add tests
Updated communication unit identifiers to use HTTP and encryption type to use `PEM`. Renamed sensor types in messaging enums for clarity and added test cases for HTTP recipient update. Changes are to aid development and testing. The reason for the change from P12 to PEM in communication unit is currently not specified. More information might be required here.