Remove outdated test fixtures and update onboarding
This commit removes outdated test fixtures related to message sending and recipient in MQTT protocol. These were too specific and had hardcoded QA environments and registration codes.

In its place, the 'onboard_response_update.py' is renamed to 'update_http_onboard_responses.py' and updated to use function 'onboard_communication_unit'. This is to align with the practice of using widely generic onboarding function which will provide more flexibility for testing.

Importantly, the new testing functions are currently marked to be skipped unless the registration code is changed. This is to avoid breaking the tests with invalid registration codes. It should be updated with valid codes when used.