Restrict parallel jobs in GitHub workflow
This commit adds 'max-parallel: 1' to the action strategy matrix in the Python-Package GitHub workflow. This change ensures that jobs in the workflow run sequentially instead of concurrently, preventing potential race conditions and conflicts in the testing environment. This helps to maintain the integrity of our test results and also keeps resource usage in check.