Refactor agrirouter to src in all imports
Removed agrirouter module and replaced it with the src module in import statements across the codebase. This change was made to better align the module names with its functions, and improve readability and organization of code structure. The change also covers respective adaptations in the file system and addition of new '__init__.py' files for the package structure. Ensured that all existing references to the old 'agrirouter' name are updated to 'src'.