Feature/move enums to api package (#48)
* Refactor enums import from messaging to api

The enums previously located in the messaging module have been moved to the api module. The import statements across the codebase have been updated accordingly. This change was done to improve the organization of the code and for better import usage.

* Remove redundant enums file in messaging module

The file src/messaging/enums.py was removed because it was redundant and no longer needed in the codebase. This cleanup contributes to the overall maintainability of the project.

* Refactor AbstractService class location

The AbstractService class was moved from messaging.py to commons.py. This refactor improves the readability and structure of the code by placing the AbstractService class in a more suitable location. The logging import was also removed from messaging.py since it is now used in commons.py.

* Update python compatibility and improve code organization

Updated the project to support Python version 3.12 and upgraded the protobuf package to version 3.20.3. Additionally, code related to environment setup, testing, and its handling in various services has been refactored to improve code organization and clarity. Removed the enumeration for environments, replaced with environment instances instead. Test cases were also adjusted accordingly.

* Update Python version compatibility

This commit modifies the project's Python version range to support from Python 3.7 through 3.11. Python 3.12 support has been removed and Python 3.7 has been added for compatibility. Changes were reflected in the README file, package metadata, and GitHub workflow configuration files.

* Add importlib-metadata to requirements

This commit adds the importlib-metadata library to the requirements.txt file. The added library is version 5.0 or earlier, ensuring compatibility with the current system.